Extract tests for TypesTokenParser into separate test case
